convert bitfields to be unsigned types
- avoid clang warnings on sign changes from 1 to -1
- convert also to smaller main type size when useful

+++ src/app_cache.c
@@ -26,7 +26,7 @@
 
 struct cache_ctx {
 	unsigned short solver_flags;
-	int add_dependencies : 1;
+	unsigned short add_dependencies : 1;
 };
 
 #define CACHE_OPTIONS(OPT) \
